File parsing is successful
